Access feedback & jsdoc fixes
